backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pdfplumber==0.10.3
pdf2image==1.16.3
Pillow==10.1.0
sklearn-crfsuite==0.3.6
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
joblib==1.3.2
python-dotenv==1.0.0

# Authentication
PyJWT==2.10.1
bcrypt==5.0.0

# Testing
pytest==8.4.2
pytest-cov==7.0.0
pytest-mock==3.12.0
rich==14.2.0